+++ src/terminal.go
@@ -57,7 +57,7 @@ var actionTypeRegex *regexp.Regexp
 const clearCode string = "\x1b[2J"
 
 func init() {
-	placeholder = regexp.MustCompile(`\\?(?:{[+sf]*[0-9,-.]*}|{q}|{fzf:(?:query|action)}|{\+?f?nf?})`)
+	placeholder = regexp.MustCompile(`\\?(?:{[+sf]*[0-9,-.]*}|{q}|{fzf:(?:query|action|prompt)}|{\+?f?nf?})`)
 	whiteSuffix = regexp.MustCompile(`\s*$`)
 	offsetComponentRegex = regexp.MustCompile(`([+-][0-9]+)|(-?/[1-9][0-9]*)`)
 	offsetTrimCharsRegex = regexp.MustCompile(`[^0-9/+-]`)
@@ -183,6 +183,7 @@ type Terminal struct {
 	separator          labelPrinter
 	separatorLen       int
 	spinner            []string
+	promptString       string
 	prompt             func()
 	promptLen          int
 	borderLabel        labelPrinter
@@ -287,6 +288,7 @@ type Terminal struct {
 	executing          *util.AtomicBool
 	termSize           tui.TermSize
 	lastAction         actionType
+	lastFocus          int32
 }
 
 type selectedItem struct {
@@ -443,6 +445,8 @@ const (
 	actRebind
 	actBecome
 	actResponse
+	actShowHeader
+	actHideHeader
 )
 
 func processExecution(action actionType) bool {
@@ -583,10 +587,14 @@ func trimQuery(query string) []rune {
 	return []rune(strings.Replace(query, "\t", " ", -1))
 }
 
-func hasPreviewAction(opts *Options) bool {
+func mayTriggerPreview(opts *Options) bool {
+	if opts.ListenAddr != nil {
+		return true
+	}
 	for _, actions := range opts.Keymap {
 		for _, action := range actions {
-			if action.t == actPreview || action.t == actChangePreview {
+			switch action.t {
+			case actPreview, actChangePreview, actTransform:
 				return true
 			}
 		}
@@ -625,8 +633,11 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		delay = initialDelay
 	}
 	var previewBox *util.EventBox
-	// We need to start previewer if HTTP server is enabled even when --preview option is not specified
-	if len(opts.Preview.command) > 0 || hasPreviewAction(opts) || opts.ListenAddr != nil {
+	// We need to start the previewer even when --preview option is not specified
+	// * if HTTP server is enabled
+	// * if 'preview' or 'change-preview' action is bound to a key
+	// * if 'transform' action is bound to a key
+	if len(opts.Preview.command) > 0 || mayTriggerPreview(opts) {
 		previewBox = util.NewEventBox()
 	}
 	var renderer tui.Renderer
@@ -670,6 +681,7 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		infoSep:            opts.InfoSep,
 		separator:          nil,
 		spinner:            makeSpinner(opts.Unicode),
+		promptString:       opts.Prompt,
 		queryLen:           [2]int{0, 0},
 		layout:             opts.Layout,
 		fullscreen:         fullscreen,
@@ -749,7 +761,8 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		tui:                renderer,
 		initFunc:           func() { renderer.Init() },
 		executing:          util.NewAtomicBool(false),
-		lastAction:         actStart}
+		lastAction:         actStart,
+		lastFocus:          minItem.Index()}
 	t.prompt, t.promptLen = t.parsePrompt(opts.Prompt)
 	t.pointer, t.pointerLen = t.processTabs([]rune(opts.Pointer), 0)
 	t.marker, t.markerLen = t.processTabs([]rune(opts.Marker), 0)
@@ -2354,7 +2367,7 @@ func parsePlaceholder(match string) (bool, string, placeholderFlags) {
 	}
 
 	if strings.HasPrefix(match, "{fzf:") {
-		// Both {fzf:query} and {fzf:action} are not determined by the current item
+		// {fzf:*} are not determined by the current item
 		flags.forceUpdate = true
 		return false, match, flags
 	}
@@ -2421,9 +2434,30 @@ func cleanTemporaryFiles() {
 	activeTempFiles = []string{}
 }
 
+type replacePlaceholderParams struct {
+	template   string
+	stripAnsi  bool
+	delimiter  Delimiter
+	printsep   string
+	forcePlus  bool
+	query      string
+	allItems   []*Item
+	lastAction actionType
+	prompt     string
+}
+
 func (t *Terminal) replacePlaceholder(template string, forcePlus bool, input string, list []*Item) string {
-	return replacePlaceholder(
-		template, t.ansi, t.delimiter, t.printsep, forcePlus, input, list, t.lastAction)
+	return replacePlaceholder(replacePlaceholderParams{
+		template:   template,
+		stripAnsi:  t.ansi,
+		delimiter:  t.delimiter,
+		printsep:   t.printsep,
+		forcePlus:  forcePlus,
+		query:      input,
+		allItems:   list,
+		lastAction: t.lastAction,
+		prompt:     t.promptString,
+	})
 }
 
 func (t *Terminal) evaluateScrollOffset() int {
@@ -2461,9 +2495,9 @@ func (t *Terminal) evaluateScrollOffset() int {
 	return util.Max(0, base)
 }
 
-func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, printsep string, forcePlus bool, query string, allItems []*Item, lastAction actionType) string {
-	current := allItems[:1]
-	selected := allItems[1:]
+func replacePlaceholder(params replacePlaceholderParams) string {
+	current := params.allItems[:1]
+	selected := params.allItems[1:]
 	if current[0] == nil {
 		current = []*Item{}
 	}
@@ -2472,7 +2506,7 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, pr
 	}
 
 	// replace placeholders one by one
-	return placeholder.ReplaceAllStringFunc(template, func(match string) string {
+	return placeholder.ReplaceAllStringFunc(params.template, func(match string) string {
 		escaped, match, flags := parsePlaceholder(match)
 
 		// this function implements the effects a placeholder has on items
@@ -2482,17 +2516,8 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, pr
 		switch {
 		case escaped:
 			return match
-		case match == "{fzf:action}":
-			name := ""
-			for i, r := range lastAction.String()[3:] {
-				if i > 0 && r >= 'A' && r <= 'Z' {
-					name += "-"
-				}
-				name += string(r)
-			}
-			return strings.ToLower(name)
 		case match == "{q}" || match == "{fzf:query}":
-			return quoteEntry(query)
+			return quoteEntry(params.query)
 		case match == "{}":
 			replace = func(item *Item) string {
 				switch {
@@ -2503,11 +2528,22 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, pr
 					}
 					return strconv.Itoa(n)
 				case flags.file:
-					return item.AsString(stripAnsi)
+					return item.AsString(params.stripAnsi)
 				default:
-					return quoteEntry(item.AsString(stripAnsi))
+					return quoteEntry(item.AsString(params.stripAnsi))
 				}
 			}
+		case match == "{fzf:action}":
+			name := ""
+			for i, r := range params.lastAction.String()[3:] {
+				if i > 0 && r >= 'A' && r <= 'Z' {
+					name += "-"
+				}
+				name += string(r)
+			}
+			return strings.ToLower(name)
+		case match == "{fzf:prompt}":
+			return quoteEntry(params.prompt)
 		default:
 			// token type and also failover (below)
 			rangeExpressions := strings.Split(match[1:len(match)-1], ",")
@@ -2522,15 +2558,15 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, pr
 			}
 
 			replace = func(item *Item) string {
-				tokens := Tokenize(item.AsString(stripAnsi), delimiter)
+				tokens := Tokenize(item.AsString(params.stripAnsi), params.delimiter)
 				trans := Transform(tokens, ranges)
 				str := joinTokens(trans)
 
 				// trim the last delimiter
-				if delimiter.str != nil {
-					str = strings.TrimSuffix(str, *delimiter.str)
-				} else if delimiter.regex != nil {
-					delims := delimiter.regex.FindAllStringIndex(str, -1)
+				if params.delimiter.str != nil {
+					str = strings.TrimSuffix(str, *params.delimiter.str)
+				} else if params.delimiter.regex != nil {
+					delims := params.delimiter.regex.FindAllStringIndex(str, -1)
 					// make sure the delimiter is at the very end of the string
 					if len(delims) > 0 && delims[len(delims)-1][1] == len(str) {
 						str = str[:delims[len(delims)-1][0]]
@@ -2550,7 +2586,7 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, pr
 		// apply 'replace' function over proper set of items and return result
 
 		items := current
-		if flags.plus || forcePlus {
+		if flags.plus || params.forcePlus {
 			items = selected
 		}
 		replacements := make([]string, len(items))
@@ -2560,7 +2596,7 @@ func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, pr
 		}
 
 		if flags.file {
-			return writeTemporaryFile(replacements, printsep)
+			return writeTemporaryFile(replacements, params.printsep)
 		}
 		return strings.Join(replacements, " ")
 	})
@@ -2735,6 +2771,13 @@ func (t *Terminal) pwindowSize() tui.TermSize {
 		size.PxHeight = size.Lines * t.termSize.PxHeight / t.termSize.Lines
 	}
 	return size
+}
+
+func (t *Terminal) currentIndex() int32 {
+	if currentItem := t.currentItem(); currentItem != nil {
+		return currentItem.Index()
+	}
+	return minItem.Index()
 }
 
 // Loop is called to start Terminal I/O
@@ -3021,17 +3064,14 @@ func (t *Terminal) Loop() {
 						t.printInfo()
 					case reqList:
 						t.printList()
-						var currentIndex int32 = minItem.Index()
-						currentItem := t.currentItem()
-						if currentItem != nil {
-							currentIndex = currentItem.Index()
-						}
+						currentIndex := t.currentIndex()
 						focusChanged := focusedIndex != currentIndex
 						if focusChanged && t.track == trackCurrent {
 							t.track = trackDisabled
 							t.printInfo()
 						}
-						if onFocus, prs := t.keymap[tui.Focus.AsEvent()]; prs && focusChanged {
+						if onFocus, prs := t.keymap[tui.Focus.AsEvent()]; prs && focusChanged && currentIndex != t.lastFocus {
+							t.lastFocus = focusedIndex
 							t.serverInputChan <- onFocus
 						}
 						if focusChanged || version != t.version {
@@ -3221,10 +3261,19 @@ func (t *Terminal) Loop() {
 		}
 
 		var doAction func(*action) bool
-		doActions := func(actions []*action) bool {
+		var doActions func(actions []*action) bool
+		doActions = func(actions []*action) bool {
+			currentIndex := t.currentIndex()
 			for _, action := range actions {
 				if !doAction(action) {
 					return false
+				}
+			}
+
+			if onFocus, prs := t.keymap[tui.Focus.AsEvent()]; prs {
+				if newIndex := t.currentIndex(); newIndex != currentIndex {
+					t.lastFocus = newIndex
+					return doActions(onFocus)
 				}
 			}
 			return true
@@ -3302,6 +3351,7 @@ func (t *Terminal) Loop() {
 				}
 			case actTransformPrompt:
 				prompt := t.executeCommand(a.a, false, true, true, true)
+				t.promptString = prompt
 				t.prompt, t.promptLen = t.parsePrompt(prompt)
 				req(reqPrompt)
 			case actTransformQuery:
@@ -3395,6 +3445,7 @@ func (t *Terminal) Loop() {
 					req(reqRedrawPreviewLabel)
 				}
 			case actChangePrompt:
+				t.promptString = a.a
 				t.prompt, t.promptLen = t.parsePrompt(a.a)
 				req(reqPrompt)
 			case actPreview:
@@ -3673,6 +3724,12 @@ func (t *Terminal) Loop() {
 					t.track = trackEnabled
 				}
 				req(reqInfo)
+			case actShowHeader:
+				t.headerVisible = true
+				req(reqList, reqInfo, reqPrompt, reqHeader)
+			case actHideHeader:
+				t.headerVisible = false
+				req(reqList, reqInfo, reqPrompt, reqHeader)
 			case actToggleHeader:
 				t.headerVisible = !t.headerVisible
 				req(reqList, reqInfo, reqPrompt, reqHeader)