+++ CHANGELOG.md
@@ -17,8 +17,9 @@ CHANGELOG
         --bind 'focus:transform:[[ -n {} ]] && exit; [[ {fzf:action} =~ up$ ]] && echo up || echo down'
   ```
 - Added placeholder expressions
-    - `{fzf:action}` - the name of the last action performed
-    - `{fzf:query}` - synonym for `{q}`
+    - `{fzf:action}` - The name of the last action performed
+    - `{fzf:prompt}` - Prompt string (including ANSI color codes)
+    - `{fzf:query}` - Synonym for `{q}`
 - Added support for negative height
   ```sh
   # Terminal height minus 1, so you can still see the command line
@@ -35,6 +36,7 @@ CHANGELOG
     ```sh
     export FZF_CTRL_R_OPTS='--bind "enter:become:if [[ -n {} ]]; then echo {}; else echo {q}; fi"'
     ```
+- Added `show-header` and `hide-header` actions
 - Bug fixes
 
 0.44.1